+++ handler/updateOpening.go
@@ -20,12 +20,12 @@ func UpdateOpeningHandler(ctx *gin.Context) {
 	}
 
 	request := struct {
-		Role     string `json:"role"`
-		Company  string `json:"company"`
-		Location string `json:"location"`
-		Remote   *bool  `json:"remote"`
-		Link     string `json:"link"`
-		Salary   int64  `json:"salary"`
+		Role     string  `json:"role"`
+		Company  string  `json:"company"`
+		Location string  `json:"location"`
+		Remote   *bool   `json:"remote"`
+		Link     string  `json:"link"`
+		Salary   float64 `json:"salary"`
 	}{}
 
 	if err := ctx.ShouldBindJSON(&request); err != nil {
